requirements.txt
# Core image processing
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-image>=0.19.0

# AI background removal
rembg>=2.0.50
torch>=1.9.0
torchvision>=0.10.0

# Development and testing
pytest>=7.0.0
matplotlib>=3.5.0
requests>=2.25.0

# Optional: GUI support (for future phases)
#tkinter  # Usually comes with Python
# pyqt5>=5.15.0  # Uncomment if you want Qt GUI later